moved READY_STATUS const to outer constants.py file
